refactor(direct-control): extract capability catalog owner
Refs: openspec/changes/civ7-support-direct-control-modularization

Move static catalog construction, runtime inspection catalog construction, official-resource capability scanning, sorting/deduplication, and private catalog helpers into src/catalog/capabilities.ts while keeping public facade exports and TypeBox schemas in index.ts.

Runtime root inspection is injected from the facade to avoid executable back-imports. Catalog contents, default roots/GameInfo tables, scan filtering, generated-output-as-evidence policy, and runtime catalog behavior are preserved. Operation/proof telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, schema/type/procedure-core ownership, and Task 2.9.4 matrix-row acceptance remain pending.

Verification:
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- git diff --check
- capability catalog ownership/no-cycle scans

Local package/source relocation proof only; no runtime, live-game, AI/hotseat product-path, telemetry, CLI semantic, or Effect/oRPC proof claimed.
