+++ packages/civ7-direct-control/src/index.ts
@@ -33,6 +33,11 @@ import {
   type FreshLogMarkerProof,
 } from "./proof/log-markers.js";
 import {
+  createStaticCiv7CapabilityCatalog as createStaticCiv7CapabilityCatalogFromModule,
+  generateCiv7CapabilityCatalog as generateCiv7CapabilityCatalogFromModule,
+  loadCiv7OfficialResourceCapabilities,
+} from "./catalog/capabilities.js";
+import {
   appUiSnapshotFromCommandResult,
   buildAppUiSnapshotCommand,
   getCiv7AppUiSnapshot as getCiv7AppUiSnapshotFromModule,
@@ -110,6 +115,7 @@ export type {
   FileSnapshot,
   FreshLogMarkerProof,
 } from "./proof/log-markers.js";
+export { loadCiv7OfficialResourceCapabilities } from "./catalog/capabilities.js";
 
 export const DEFAULT_CIV7_TUNER_HOST = "127.0.0.1";
 export const DEFAULT_CIV7_TUNER_PORT = 4318;
@@ -3963,103 +3969,20 @@ export async function requestCiv7UnitTargetAction(
 }
 
 export function createStaticCiv7CapabilityCatalog(): Civ7CapabilityCatalog {
-  return {
-    generatedAt: new Date().toISOString(),
-    source: "static",
-    version: "direct-control-v1",
-    entries: [
-      ...STATIC_CIV7_CAPABILITY_ENTRIES,
-      ...DEFAULT_CIV7_GAMEINFO_TABLES.map((table): Civ7CapabilityCatalogEntry => ({
-        id: `gameinfo.${table}`,
-        name: table,
-        role: "tuner",
-        kind: "gameinfo-table",
-        owner: "@civ7/direct-control",
-        risk: "read",
-        provenance: ["DEFAULT_CIV7_GAMEINFO_TABLES", "capability-inventory"],
-        confidence: "source",
-        description: `Targeted GameInfo.${table} read surface.`,
-      })),
-    ],
-  };
+  return createStaticCiv7CapabilityCatalogFromModule({
+    gameinfoTables: DEFAULT_CIV7_GAMEINFO_TABLES,
+  });
 }
 
 export async function generateCiv7CapabilityCatalog(
   options: Civ7CapabilityCatalogOptions = {},
 ): Promise<Civ7CapabilityCatalog> {
-  const includeStatic = options.includeStatic !== false;
-  const includeRuntime = options.includeRuntime !== false;
-  const entries: Civ7CapabilityCatalogEntry[] = includeStatic
-    ? [...createStaticCiv7CapabilityCatalog().entries]
-    : [];
-  if (includeRuntime) {
-    const [appUi, tuner] = await Promise.all([
-      inspectCiv7Root({
-        state: { role: "app-ui" },
-        roots: options.appUiRoots ?? DEFAULT_CIV7_CAPABILITY_APP_UI_ROOTS,
-        maxRoots: 32,
-        maxKeys: 128,
-        maxMethods: 128,
-        includeSignatures: false,
-      }, options),
-      inspectCiv7Root({
-        state: { role: "tuner" },
-        roots: options.tunerRoots ?? DEFAULT_CIV7_CAPABILITY_TUNER_ROOTS,
-        maxRoots: 32,
-        maxKeys: 128,
-        maxMethods: 128,
-        includeSignatures: false,
-      }, options),
-    ]);
-    entries.push(...capabilityEntriesFromInspection(appUi, "app-ui"));
-    entries.push(...capabilityEntriesFromInspection(tuner, "tuner"));
-  }
-  return dedupeCapabilityCatalog({
-    generatedAt: new Date().toISOString(),
-    source: includeStatic && includeRuntime ? "merged" : includeRuntime ? "runtime" : "static",
-    version: "direct-control-v1",
-    entries,
+  return await generateCiv7CapabilityCatalogFromModule(options, {
+    appUiRoots: DEFAULT_CIV7_CAPABILITY_APP_UI_ROOTS,
+    gameinfoTables: DEFAULT_CIV7_GAMEINFO_TABLES,
+    inspectRoot: inspectCiv7Root,
+    tunerRoots: DEFAULT_CIV7_CAPABILITY_TUNER_ROOTS,
   });
-}
-
-export async function loadCiv7OfficialResourceCapabilities(options: {
-  resourcesRoot: string;
-  maxFiles?: number;
-}): Promise<ReadonlyArray<Civ7CapabilityCatalogEntry>> {
-  const maxFiles = options.maxFiles ?? 2_000;
-  const files = await listFiles(options.resourcesRoot, maxFiles);
-  const patterns: ReadonlyArray<Readonly<{ id: string; label: string; pattern: RegExp; risk: Civ7CapabilityCatalogEntry["risk"] }>> = [
-    { id: "official.Network.restartGame", label: "Network.restartGame", pattern: /Network\.restartGame/g, risk: "medium" },
-    { id: "official.UI.notifyUIReady", label: "UI.notifyUIReady", pattern: /UI\.notifyUIReady/g, risk: "medium" },
-    { id: "official.Autoplay", label: "Autoplay setters", pattern: /Autoplay\.set(?:Active|Turns|Pause|ObserveAsPlayer|ReturnAsPlayer)/g, risk: "medium" },
-    { id: "official.GameContext.turn", label: "GameContext turn completion", pattern: /GameContext\.(?:sendTurnComplete|sendUnreadyTurn|hasSentTurnComplete)/g, risk: "medium" },
-    { id: "official.Visibility.revealAllPlots", label: "Visibility.revealAllPlots", pattern: /Visibility\.revealAllPlots/g, risk: "high" },
-    { id: "official.Game.UnitOperations", label: "Game.UnitOperations", pattern: /Game\.UnitOperations\.(?:canStart|sendRequest)/g, risk: "medium" },
-    { id: "official.Game.CityOperations", label: "Game.CityOperations", pattern: /Game\.CityOperations\.(?:canStart|sendRequest)/g, risk: "medium" },
-    { id: "official.Game.PlayerOperations", label: "Game.PlayerOperations", pattern: /Game\.PlayerOperations\.(?:canStart|sendRequest)/g, risk: "medium" },
-  ];
-  const found = new Map<string, Civ7CapabilityCatalogEntry>();
-  for (const file of files) {
-    if (!/\.(js|ts|xml|sql)$/i.test(file)) continue;
-    const text = await readFile(file, "utf8").catch(() => "");
-    for (const pattern of patterns) {
-      if (!pattern.pattern.test(text)) continue;
-      pattern.pattern.lastIndex = 0;
-      found.set(pattern.id, {
-        id: pattern.id,
-        name: pattern.label,
-        role: pattern.id.includes("Network") || pattern.id.includes("UI.") || pattern.id.includes("GameContext")
-          ? "app-ui"
-          : "tuner",
-        kind: "method",
-        owner: "official-resources",
-        risk: pattern.risk,
-        provenance: [file],
-        confidence: "source",
-      });
-    }
-  }
-  return Array.from(found.values());
 }
 
 async function openCiv7TunerSocket(options: {
@@ -5143,160 +5066,6 @@ function narrativeChoiceRequestSource(): string {
     };`;
 }
 
-const STATIC_CIV7_CAPABILITY_ENTRIES: ReadonlyArray<Civ7CapabilityCatalogEntry> = [
-  {
-    id: "wrapper.restart-begin",
-    name: "Restart and Begin",
-    role: "app-ui",
-    kind: "action-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "medium",
-    provenance: ["restartCiv7GameAndBegin", "live-owner-proof"],
-    wrapper: "restartCiv7GameAndBegin",
-    confidence: "recorded-live-proof",
-    description: "Runs Network.restartGame, follows native Begin Game readiness, and waits for Tuner readiness.",
-  },
-  {
-    id: "wrapper.playable-status",
-    name: "Playable Status",
-    role: "shared",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["getCiv7AppUiSnapshot", "checkCiv7TunerHealth"],
-    wrapper: "getCiv7PlayableStatus",
-    confidence: "runtime",
-  },
-  {
-    id: "wrapper.setup-snapshot",
-    name: "Setup Snapshot",
-    role: "app-ui",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["Configuration", "GameSetup", "Database", "studio-run-in-game"],
-    wrapper: "getCiv7SetupSnapshot|getCiv7SetupMapRows",
-    confidence: "runtime",
-    description: "Reads Civ7 setup phase, setup parameters, and frontend map-script row visibility.",
-  },
-  {
-    id: "wrapper.setup-start",
-    name: "Single-player Setup and Start",
-    role: "app-ui",
-    kind: "action-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "high",
-    provenance: ["Configuration.editMap", "Network.hostGame", "studio-run-in-game"],
-    wrapper: "prepareCiv7SinglePlayerSetup|startPreparedCiv7SinglePlayerGame|runCiv7SinglePlayerFromSetup",
-    confidence: "source",
-    description: "Applies map script, map size, and map seed through App UI setup APIs, then starts a prepared single-player game.",
-  },
-  {
-    id: "wrapper.map-summary",
-    name: "Map Summary",
-    role: "tuner",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["GameplayMap", "Game"],
-    wrapper: "getCiv7MapSummary",
-    confidence: "recorded-live-proof",
-  },
-  {
-    id: "wrapper.plot-grid",
-    name: "Plot and Grid Snapshots",
-    role: "tuner",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["GameplayMap", "Visibility", "MapUnits", "MapCities"],
-    wrapper: "getCiv7PlotSnapshot|getCiv7MapGrid|getCiv7FullMapGrid",
-    confidence: "recorded-live-proof",
-  },
-  {
-    id: "wrapper.resource-placement-feasibility",
-    name: "Resource Placement Feasibility",
-    role: "tuner",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["ResourceBuilder.canHaveResource", "ResourceBuilder.getBestMapResourceCuts", "ResourceBuilder.getResourceCounts", "GameInfo.Resources", "GameplayMap"],
-    wrapper: "getCiv7ResourcePlacementFeasibility|getCiv7ResourceBuilderDiagnostics",
-    confidence: "source",
-    description: "Reads bounded per-cell resource feasibility and ResourceBuilder cut/count diagnostics for parity/source-authority diagnostics without mutating the map.",
-  },
-  {
-    id: "wrapper.feature-placement-feasibility",
-    name: "Feature Placement Feasibility",
-    role: "tuner",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["TerrainBuilder.canHaveFeature", "GameplayMap"],
-    wrapper: "getCiv7FeaturePlacementFeasibility",
-    confidence: "source",
-    description: "Reads bounded per-cell feature feasibility for parity/source-authority diagnostics without mutating the map.",
-  },
-  {
-    id: "wrapper.autoplay",
-    name: "Autoplay Control",
-    role: "app-ui",
-    kind: "action-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "medium",
-    provenance: ["Autoplay"],
-    wrapper: "configureCiv7Autoplay|startCiv7Autoplay|stopCiv7Autoplay",
-    confidence: "source",
-  },
-  {
-    id: "wrapper.target-candidates",
-    name: "Target Candidates",
-    role: "app-ui",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["Players", "Cities", "Units", "GameInfo"],
-    wrapper: "getCiv7TargetCandidates",
-    confidence: "runtime",
-    description: "Ranks candidate other-owner contacts from runtime city/unit summaries and a supplied formation origin.",
-  },
-  {
-    id: "wrapper.battlefield-scan",
-    name: "Battlefield Scan",
-    role: "app-ui",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["Players", "Cities", "Units", "GameInfo"],
-    wrapper: "getCiv7BattlefieldScan",
-    confidence: "runtime",
-    description: "Summarizes nearby units, cities, owner pressure, and tactical points of interest around supplied origins.",
-  },
-  {
-    id: "wrapper.destination-analysis",
-    name: "Destination Analysis",
-    role: "app-ui",
-    kind: "read-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "read",
-    provenance: ["Players", "Cities", "Units", "GameplayMap", "GameInfo"],
-    wrapper: "getCiv7DestinationAnalysis",
-    confidence: "runtime",
-    description: "Summarizes pressure near an intended destination and along a cheap straight-line corridor without issuing movement.",
-  },
-  {
-    id: "wrapper.operations",
-    name: "Validator-backed gameplay operations",
-    role: "tuner",
-    kind: "action-wrapper",
-    owner: "@civ7/direct-control",
-    risk: "medium",
-    provenance: ["Game.UnitOperations", "Game.UnitCommands", "Game.CityOperations", "Game.CityCommands", "Game.PlayerOperations"],
-    wrapper: "canStart*/request*",
-    confidence: "recorded-live-proof",
-  },
-];
-
 function jsonPayloadFromCommandResult<T extends object>(result: Civ7CommandResult, label: string): T {
   try {
     const payload = JSON.parse(result.output[0] ?? "{}") as T;
@@ -6281,68 +6050,6 @@ function validateOperationInput(family: Civ7OperationFamily, input: Civ7Operatio
   if (family === "player-operation" && !("playerId" in input)) {
     throw new Civ7DirectControlError("command-failed", "player-operation requires playerId");
   }
-}
-
-function capabilityEntriesFromInspection(
-  inspection: Civ7RootInspectionResult,
-  role: "app-ui" | "tuner",
-): Civ7CapabilityCatalogEntry[] {
-  const entries: Civ7CapabilityCatalogEntry[] = [];
-  for (const root of inspection.roots) {
-    entries.push({
-      id: `${role}.root.${root.name}`,
-      name: root.name,
-      role,
-      kind: "root",
-      owner: "runtime",
-      risk: "read",
-      provenance: [`${inspection.state.name}:${root.name}`],
-      state: inspection.state.name,
-      root: root.name,
-      confidence: "runtime",
-    });
-    for (const method of root.methods) {
-      entries.push({
-        id: `${role}.method.${root.name}.${method.name}`,
-        name: `${root.name}.${method.name}`,
-        role,
-        kind: "method",
-        owner: "runtime",
-        risk: method.name.startsWith("get") || method.name.startsWith("is") || method.name.startsWith("has") ? "read" : "medium",
-        provenance: [`${inspection.state.name}:${root.name}.${method.name}`],
-        state: inspection.state.name,
-        root: root.name,
-        method: method.name,
-        confidence: "runtime",
-      });
-    }
-  }
-  return entries;
-}
-
-function dedupeCapabilityCatalog(catalog: Civ7CapabilityCatalog): Civ7CapabilityCatalog {
-  const byId = new Map<string, Civ7CapabilityCatalogEntry>();
-  for (const entry of catalog.entries) {
-    const existing = byId.get(entry.id);
-    byId.set(entry.id, existing ? { ...existing, provenance: Array.from(new Set([...existing.provenance, ...entry.provenance])) } : entry);
-  }
-  return { ...catalog, entries: Array.from(byId.values()).sort((a, b) => a.id.localeCompare(b.id)) };
-}
-
-async function listFiles(root: string, maxFiles: number): Promise<string[]> {
-  const out: string[] = [];
-  async function visit(dir: string): Promise<void> {
-    if (out.length >= maxFiles) return;
-    const entries = await readdir(dir, { withFileTypes: true }).catch(() => []);
-    for (const entry of entries) {
-      if (out.length >= maxFiles) return;
-      const path = join(dir, entry.name);
-      if (entry.isDirectory()) await visit(path);
-      else if (entry.isFile()) out.push(path);
-    }
-  }
-  await visit(root);
-  return out;
 }
 
 function jsLiteral(value: unknown): string {
